% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/best_conditions.R
\docType{data}
\name{best_conditions}
\alias{best_conditions}
\title{Best surfing conditions by spot}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 4698 rows and 13 columns.
}
\source{
https://surf-forecast.com/.
}
\usage{
best_conditions
}
\description{
Best surfing conditions by spot
}
\keyword{datasets}
