% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/country_spot.R
\docType{data}
\name{country_spot}
\alias{country_spot}
\title{An internal dataset}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 7884 rows and 4 columns.
}
\source{
https://surf-forecast.com/.
}
\usage{
country_spot
}
\description{
An internal dataset
}
\keyword{datasets}
