% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Run the app to find the best surf spots for a given time of the year.}
\usage{
run_app()
}
\value{
Called for its side effect.
}
\description{
Run the app to find the best surf spots for a given time of the year.
}
\examples{
if (interactive()) run_app()
}
