% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surfcast-package.R
\docType{package}
\name{surfcast-package}
\alias{surfcast}
\alias{surfcast-package}
\title{surfcast: Find the Best Surfing Spot for a Given Time of the Year}
\description{
Find the best surfing spot for a given time of the year.
}
\author{
\strong{Maintainer}: Mauro Lepore \email{maurolepore@gmail.com} (\href{https://orcid.org/0000-0002-1986-7988}{ORCID}) [contractor]

}
\keyword{internal}
